don't overwrite database.yml if it already exists (current implementation does not work with JRuby/JDBC)

+++ lib/language_pack/ruby.rb
@@ -504,6 +504,7 @@ ERROR
   def create_database_yml
     log("create_database_yml") do
       return unless File.directory?("config")
+      return if File.exist?("config/database.yml")
       topic("Writing config/database.yml to read from DATABASE_URL")
       File.open("config/database.yml", "w") do |file|
         file.puts <<-DATABASE_YML
